Exposing a method in the BaseView for clearing event handlers. (#64)
* Exposing a method in the BaseView for clearing event handlers.

The instance of event actions is now preserved as a private property so we can use it when
calling to clearEvents.

* Using the eventListenerName properly for storing the event listener map.

We were using the eventName (e.g. 'click') instead of the
eventListenerName (e.g. onCloseClick), which caused that we were not
properly storing the event handlers in the eventMap. As a consequence
we were not able to dettach in every view that had more than one event
handler active attach to the same action (e.g. 2 clicks in the same
view, a pretty standard scenario)